Refactor nmcli interactions into separate modules
This commit reorganizes the code by separating NetworkManager command-line interface interactions into distinct modules: `nm_ports` for port management and `nm_query` for querying connection names. This improves code organization and maintainability.
